+++ src/mailman/docs/NEWS.rst
@@ -29,6 +29,7 @@ REST
    member records.  Arguments are `subscriber` (email address to search for -
    required), `fqdn_listname` (optional), and `role` (i.e. MemberRole -
    optional).  (LP: #799612)
+ * Fixed /lists/<fqdn_listname>/<role>/<email> (LP: #825570)
 
 Commands
 --------